// Frontend/src/app/api/model/createTodoDto.ts
/**
 * Campus Manager
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateTodoDto { 
    name: string;
    description: string;
    date: string;
    startTime: string;
    endTime: string;
}