// Frontend/src/app/api/model/updateTodoDto.ts
/**
 * Campus Manager
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateTodoDto { 
    name?: string;
    description?: string;
    date?: string;
    startTime?: string;
    endTime?: string;
}